feat: inf loop handling
- limit max output length
- enable float max runtime

+++ eval.py
@@ -23,7 +23,7 @@ def main():
     parser.add_argument('--unzip-code', default=False, action='store_true', help='LearnUs 압축파일에서 코드 추출 여부')
     parser.add_argument('--ignore-blanks', default=True, action='store_true', help='정답 비교 시 Whitespace(\\t, \\n, 공백) 무시')
     parser.add_argument('--ignore-capitals', default=True, action='store_true', help='정답 비교 시 대/소문자 구분 안 함')
-    parser.add_argument('--timeout', type=int, default=5, help='최대 실행 시간')
+    parser.add_argument('--timeout', type=float, default=5, help='최대 실행 시간')
     parser.add_argument('--thread', type=int, default=32, help='최대 스레드 수')
 
     args = parser.parse_args()
